[#11024] DocDB: Enable -Wshorten-64-to-32 flag
Summary:
This diff enables -Wshorten-64-to-32 flag and fixes all places where this warning appear in our code.
The following strategy was used:
1) Minimize amount of casts, i.e. changing used type is more preferable.
2) For places where such cast is necessary, for instance because of 3rd party library API, use newly introduced `narrow_cast`.
`narrow_cast` expects that casted values left unchanged. I.e. if we cast 64 bit number to 32 bit number it expects that actual value fits into 32 bits.

Test Plan: Jenkins

Reviewers: mbautin, mpolitov

Reviewed By: mpolitov

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D14653

+++ src/yb/bfql/bfunc_convert.h
@@ -36,6 +36,7 @@
 #include "yb/common/ql_datatype.h"
 #include "yb/common/ql_type.h"
 
+#include "yb/gutil/casts.h"
 #include "yb/gutil/endian.h"
 #include "yb/gutil/strings/escaping.h"
 #include "yb/util/date_time.h"
@@ -469,7 +470,7 @@ CHECKED_STATUS ConvertI64ToI32(PTypePtr source, RTypePtr target) {
   if (source->IsNull()) {
     target->SetNull();
   } else {
-    target->set_int32_value(source->int64_value());
+    target->set_int32_value(static_cast<int32_t>(source->int64_value()));
   }
   return Status::OK();
 }
@@ -1193,7 +1194,7 @@ CHECKED_STATUS ConvertVarintToI32(PTypePtr source, RTypePtr target) {
     if (val < INT32_MIN || val > INT32_MAX) {
       return STATUS(QLError, "VarInt cannot be converted to int32 due to overflow");
     }
-    target->set_int32_value(val);
+    target->set_int32_value(static_cast<int32_t>(val));
   }
   return Status::OK();
 }
@@ -1274,8 +1275,8 @@ CHECKED_STATUS ConvertI64ToVarint(PTypePtr source, RTypePtr target) {
 }
 
 template<typename RTypePtr>
-CHECKED_STATUS ToInt32(int32_t val, RTypePtr target) {
-  target->set_int32_value(val);
+CHECKED_STATUS ToInt32(int64_t val, RTypePtr target) {
+  target->set_int32_value(static_cast<int32_t>(val));
   return Status::OK();
 }
